[![Mend
Renovate](https://app.renovatebot.com/images/banner.svg)](https://renovatebot.com)

This PR contains the following updates:

| Package | Change | Age | Adoption | Passing | Confidence |
|---|---|---|---|---|---|
|
[github.com/equinix/equinix-sdk-go](https://togithub.com/equinix/equinix-sdk-go)
| `v0.30.0` -> `v0.31.1` |
[![age](https://developer.mend.io/api/mc/badges/age/go/github.com%2fequinix%2fequinix-sdk-go/v0.31.1?slim=true)](https://docs.renovatebot.com/merge-confidence/)
|
[![adoption](https://developer.mend.io/api/mc/badges/adoption/go/github.com%2fequinix%2fequinix-sdk-go/v0.31.1?slim=true)](https://docs.renovatebot.com/merge-confidence/)
|
[![passing](https://developer.mend.io/api/mc/badges/compatibility/go/github.com%2fequinix%2fequinix-sdk-go/v0.30.0/v0.31.1?slim=true)](https://docs.renovatebot.com/merge-confidence/)
|
[![confidence](https://developer.mend.io/api/mc/badges/confidence/go/github.com%2fequinix%2fequinix-sdk-go/v0.30.0/v0.31.1?slim=true)](https://docs.renovatebot.com/merge-confidence/)
|

---

### Release Notes

<details>
<summary>equinix/equinix-sdk-go
(github.com/equinix/equinix-sdk-go)</summary>

###
[`v0.31.1`](https://togithub.com/equinix/equinix-sdk-go/releases/tag/v0.31.1)

[Compare
Source](https://togithub.com/equinix/equinix-sdk-go/compare/v0.31.0...v0.31.1)

#####
[0.31.1](https://togithub.com/equinix/equinix-sdk-go/compare/v0.31.0...v0.31.1)
(2024-01-19)

##### Bug Fixes

- specify correct type for interconnection speed
([#&#8203;22](https://togithub.com/equinix/equinix-sdk-go/issues/22))
([2cd2e50](https://togithub.com/equinix/equinix-sdk-go/commit/2cd2e504568ea93650c4b3c6072abb21256a0b63)),
closes
[#&#8203;16](https://togithub.com/equinix/equinix-sdk-go/issues/16)

###
[`v0.31.0`](https://togithub.com/equinix/equinix-sdk-go/releases/tag/v0.31.0)

[Compare
Source](https://togithub.com/equinix/equinix-sdk-go/compare/v0.30.0...v0.31.0)

##### Features

- API Sync by GitHub Action for
[@&#8203;ctreatma](https://togithub.com/ctreatma)
([#&#8203;10](https://togithub.com/equinix/equinix-sdk-go/issues/10))
([8096b49](https://togithub.com/equinix/equinix-sdk-go/commit/8096b497e4a321181563e730c39f7710e415ce98))

</details>

---

### Configuration

📅 **Schedule**: Branch creation - At any time (no schedule defined),
Automerge - At any time (no schedule defined).

🚦 **Automerge**: Disabled by config. Please merge this manually once you
are satisfied.

♻ **Rebasing**: Whenever PR becomes conflicted, or you tick the
rebase/retry checkbox.

🔕 **Ignore**: Close this PR and you won't be reminded about this update
again.

---

- [ ] <!-- rebase-check -->If you want to rebase/retry this PR, check
this box

---

This PR has been generated by [Mend
Renovate](https://www.mend.io/free-developer-tools/renovate/). View
repository job log
[here](https://developer.mend.io/github/equinix/metal-cli).

<!--renovate-debug:eyJjcmVhdGVkSW5WZXIiOiIzNy4xMDMuMSIsInVwZGF0ZWRJblZlciI6IjM3LjEzNS4wIiwidGFyZ2V0QnJhbmNoIjoibWFpbiJ9-->

+++ internal/virtualcircuit/retrieve.go
@@ -13,7 +13,7 @@ import (
 type vcParamOuter interface {
 	GetId() string
 	GetName() string
-	GetSpeed() int32
+	GetSpeed() int64
 	GetCreatedAt() time.Time
 }
